+++ src/Sulu/Bundle/MediaBundle/Controller/MediaStreamController.php
@@ -126,13 +126,11 @@ class MediaStreamController extends Controller
 
         $response = new BinaryFileResponse($path);
 
-        $pathInfo = pathinfo($fileName);
-
         // Prepare headers
         $disposition = $response->headers->makeDisposition(
             $dispositionType,
             $fileName,
-            $cleaner->cleanup($pathInfo['filename'], $locale) . '.' . $pathInfo['extension']
+            $this->cleanUpFileName($fileName, $locale, $fileVersion->getExtension())
         );
 
         // Set headers
@@ -181,6 +179,26 @@ class MediaStreamController extends Controller
         }
 
         return $currentFileVersion;
+    }
+
+    /**
+     * Cleaned up filename.
+     *
+     * @param string $fileName
+     * @param string $locale
+     * @param string $extension
+     *
+     * @return string
+     */
+    private function cleanUpFileName($fileName, $locale, $extension)
+    {
+        $pathInfo = pathinfo($fileName);
+        $cleanedFileName = $this->get('sulu.content.path_cleaner')->cleanup($pathInfo['filename'], $locale);
+        if ($extension) {
+            $cleanedFileName .= '.' . $extension;
+        }
+
+        return $cleanedFileName;
     }
 
     /**